requirements.txt
blinker==1.9.0
click==8.2.1
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.3.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
